Backend/restapi/requirements.txt
anyio==4.12.1
asgiref==3.7.2
attrs==21.2.0
Automat==20.2.0
aws-cfn-bootstrap @ https://s3.amazonaws.com/cloudformation-examples/aws-cfn-bootstrap-py3-latest.tar.gz
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
chevron==0.14.0
click==8.0.3
cloud-init==25.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
Cython==3.2.4
dartsclone==0.9.0
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.17
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
dnspython==2.8.0
docutils==0.19
ec2-hibinit-agent==1.0.0
exceptiongroup==1.3.1
gyp==0.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.25.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
MarkupSafe==2.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==21.3
pexpect==4.8.0
psutil==7.0.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
pygtrie==2.5.0
PyHamcrest==2.0.2
PyJWT==2.3.0
pymongo==4.15.5
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4.1
python-daemon==2.1.2
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
python-telegram-bot==20.7
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
ruff==0.14.10
SecretStorage==3.3.1
service-identity==18.1.0
shelve_cache==0.0.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.1.0
sos==4.9.2
sqlparse==0.5.5
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
typing_extensions==4.15.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
zipp==1.0.0
zope.interface==5.4.0
